do not copy the auth_router
