--- fixbackend/auth/auth_router.py
@@ -1,107 +1,0 @@
-#  Copyright (c) 2023. Some Engineering
-#  Copyright (c) 2019 François Voron
-#  This program is free software: you can redistribute it and/or modify
-#  it under the terms of the GNU Affero General Public License as published by
-#  the Free Software Foundation, either version 3 of the License, or
-#  (at your option) any later version.
-#
-#  This program is distributed in the hope that it will be useful,
-#  but WITHOUT ANY WARRANTY; without even the implied warranty of
-#  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-#  GNU Affero General Public License for more details.
-#
-#  You should have received a copy of the GNU Affero General Public License
-#  along with this program.  If not, see <http://www.gnu.org/licenses/>.
-
-from typing import Optional, Tuple
-
-from fastapi import APIRouter, Depends, HTTPException, Request, Response, status
-from fastapi.security import OAuth2PasswordRequestForm
-
-from fastapi_users.authentication import AuthenticationBackend, Authenticator, Strategy
-from fastapi_users.manager import BaseUserManager, UserManagerDependency
-from fastapi_users.openapi import OpenAPIResponseType
-from fastapi_users.router.common import ErrorCode, ErrorModel
-
-from fixbackend.auth.models import User
-from fixbackend.ids import UserId
-
-
-def get_auth_router(
-    backend: AuthenticationBackend[User, UserId],
-    get_user_manager: UserManagerDependency[User, UserId],
-    authenticator: Authenticator,
-    requires_verification: bool = False,
-) -> APIRouter:
-    """Generate a router with login/logout routes for an authentication backend."""
-    router = APIRouter()
-    get_current_user_token = authenticator.current_user_token(
-        optional=True, active=True, verified=requires_verification
-    )
-
-    login_responses: OpenAPIResponseType = {
-        status.HTTP_400_BAD_REQUEST: {
-            "model": ErrorModel,
-            "content": {
-                "application/json": {
-                    "examples": {
-                        ErrorCode.LOGIN_BAD_CREDENTIALS: {
-                            "summary": "Bad credentials or the user is inactive.",
-                            "value": {"detail": ErrorCode.LOGIN_BAD_CREDENTIALS},
-                        },
-                        ErrorCode.LOGIN_USER_NOT_VERIFIED: {
-                            "summary": "The user is not verified.",
-                            "value": {"detail": ErrorCode.LOGIN_USER_NOT_VERIFIED},
-                        },
-                    }
-                }
-            },
-        },
-        **backend.transport.get_openapi_login_responses_success(),
-    }
-
-    @router.post(
-        "/login",
-        name=f"auth:{backend.name}.login",
-        responses=login_responses,
-    )
-    async def login(
-        request: Request,
-        credentials: OAuth2PasswordRequestForm = Depends(),
-        user_manager: BaseUserManager[User, UserId] = Depends(get_user_manager),
-        strategy: Strategy[User, UserId] = Depends(backend.get_strategy),
-    ) -> Response:
-        user = await user_manager.authenticate(credentials)
-
-        if user is None or not user.is_active:
-            raise HTTPException(
-                status_code=status.HTTP_400_BAD_REQUEST,
-                detail=ErrorCode.LOGIN_BAD_CREDENTIALS,
-            )
-        if requires_verification and not user.is_verified:
-            raise HTTPException(
-                status_code=status.HTTP_400_BAD_REQUEST,
-                detail=ErrorCode.LOGIN_USER_NOT_VERIFIED,
-            )
-        response = await backend.login(strategy, user)
-        await user_manager.on_after_login(user, request, response)
-        return response
-
-    logout_responses: OpenAPIResponseType = {
-        **{status.HTTP_401_UNAUTHORIZED: {"description": "Missing token or inactive user."}},
-        **backend.transport.get_openapi_logout_responses_success(),
-    }
-
-    @router.post("/logout", name=f"auth:{backend.name}.logout", responses=logout_responses)
-    async def logout(
-        user_token: Tuple[Optional[User], Optional[str]] = Depends(get_current_user_token),
-        strategy: Strategy[User, UserId] = Depends(backend.get_strategy),
-    ) -> Response:
-        user, token = user_token
-        if user is None or token is None:
-            # no one to logout
-            return Response(status_code=status.HTTP_204_NO_CONTENT)
-
-        return await backend.logout(strategy, user, token)
-
-    return router
